chore: Rust 1.89.0 linting

+++ src/servers/api/routers/user.rs
@@ -147,21 +147,21 @@ impl UserRouter {
                 ij::Token::Totp(token) => {
                     let known_totp = authentication::totp_from_secret(two_fa_setup.value())?;
 
-                    if let Ok(valid_token) = known_totp.check_current(&token) {
-                        if valid_token {
-                            RedisTwoFASetup::delete(&state.redis, &user).await?;
-                            ModelTwoFA::insert(&state.postgres, two_fa_setup, useragent_ip, &user)
-                                .await?;
+                    if let Ok(valid_token) = known_totp.check_current(&token)
+                        && valid_token
+                    {
+                        RedisTwoFASetup::delete(&state.redis, &user).await?;
+                        ModelTwoFA::insert(&state.postgres, two_fa_setup, useragent_ip, &user)
+                            .await?;
 
-                            Email::new(
-                                &user.full_name,
-                                &user.email,
-                                EmailTemplate::TwoFAEnabled,
-                                &state.email_env,
-                            )
-                            .send();
-                            return Ok(axum::http::StatusCode::OK);
-                        }
+                        Email::new(
+                            &user.full_name,
+                            &user.email,
+                            EmailTemplate::TwoFAEnabled,
+                            &state.email_env,
+                        )
+                        .send();
+                        return Ok(axum::http::StatusCode::OK);
                     }
                 }
                 ij::Token::Backup(_) => return err(),
